Added fsl,msi property to PCI node
The association between PCI and MSI in the hardware device tree
through fsl,msi property is an artificial one and it doesn't
reflect the actual hardware.
The user can now make the association between MSI and PCI using
node-update-phandle in the hypervisor config tree. The association
in the hardware device tree will be removed.
e.g.

	pci0 {
		device = "pci0";
		map-ranges;
		dma-window = <&dw_linux1>;
		node-update-phandle {
			fsl,msi = <&msi0>;
		};
	};

If no association is present, Topaz is creating it by associating
the first MSI node found with any PCI controllers.

Signed-off-by: Diana CRACIUN <Diana.Craciun@freescale.com>