Changed the MSI subwindow position to be just after guest
The MSI subwindow is used for MSIIR translation. The hypervisor
sets a property in the guest device tree (msi-address-64) which
indicates the guest physical address of the MSIIR register. This
address must be in the next page after the end of DDR as seen by
the guest. This is because Linux is extending the DDR ATMU with
one page in order to cover the MSIIR register.

Signed-off-by: Diana CRACIUN <Diana.Craciun@freescale.com>
Signed-off-by: Laurentiu Tudor <Laurentiu.Tudor@freescale.com>