p2041rdb: Initial support for shared-mac interfaces
Added Shared-MAC support between two Linux partitions running
under HV.

Without setting the fman node in the second partition, the Fman
access to the memory will not be authorized by PAMU due to invalid
LIODN, which will result in a memory access violation.

In Shared-MAC scenario, PAMU is using Fman's LIODN in order
to authorize tranzactions for the second partition, thus the
fman node have to be included in the second partition, so that
PAMU to be configured accordingly.

Signed-off-by: Cristian Sovaiala <cristian.sovaiala@freescale.com>