p2041rdb: Added Shared-MAC interfaces
Defined interface in partition 1 to share the MAC with interface
in partition 2.
Added interface in the second partition which shares the MAC with
the interface in the first partition.

Both interfaces shared the same TX queues but have different RX
queues. The first partition initializes the MAC and the second
partition removes the fsl,fman-mac property.

Created new buffer-pool nodes for shared interfaces.

Signed-off-by: Cristian Sovaiala <cristian.sovaiala@freescale.com>