p5040: initial hv-2p scenario
There is one change to how we normaly do device assignment:
On P5040, the hypervisor is issuing a "cannot share IRQ" error.
This is because on the new SuperHydra boards the DS3232 RTC which
is also connected to external IRQ1 pin was moved from I2C bus 2
(assigned to "part1") to I2C bus 4 (assigned to "part2").
But PCIe1 (assigned to "part1") INT-B is multiplexed on IRQ1.
The hypervisor detects that IRQ1 is used in both guests thus
issuing the error above. To make it go away, simply move I2C
assigned to 2nd partition to 1st.

Signed-off-by: Laurentiu Tudor <Laurentiu.Tudor@freescale.com>