Increase guest dtb window from 72KB to 128KB
On P5040ds the memory window reserved for the guest dtb
was no longer large enough to fit the flattened device tree,
(the device tree size for part1 reached 73972 bytes)
causing the hypervisor to fail to generate de guest dtb
and thus start the partition.
Increase the dtb window from 72KB to a safe value of 128KB
for all platforms, just to be on the safe side.

Signed-off-by: Laurentiu Tudor <Laurentiu.Tudor@freescale.com>